# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Sentinel AML API
  version: 1.1.0
  description: Static API contract for the implementation in this repository. Basic
    authentication is stateless. Backend-configured users are admin, analyst and viewer;
    credentials come from environment variables and are never included here. All list
    responses mask customer identifiers and sensitive narrative fields. Numeric customer
    IDs support the authorized by-id detail endpoint. VIEWER detail views remain redacted.
    CSV and JSON batch submission returns 202 with a durable job to poll. Detection
    occurs atomically for each processed transaction. No external
    SAR filing API is implemented.
servers:
- url: http://localhost:8080/api/v1
  description: Local development; use HTTPS for deployment
security:
- basicAuth: []
paths:
...
  "/accounts":
    get:
      operationId: listAccounts
      summary: List accounts
      tags:
      - Accounts
      description: 'Allowed roles: ADMIN, ANALYST, VIEWER.'
      x-roles:
      - ADMIN
      - ANALYST
      - VIEWER
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PageAccountResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          minimum: 1
          default: 20
      - name: sort
        in: query
        schema:
          type: array
          items:
            type: string
        style: form
        explode: true
        description: Repeat for multiple sorts, e.g. sort=riskScore,desc.
    post:
      operationId: createAccount
      summary: Create or ingest account
      tags:
      - Accounts
      description: 'Allowed roles: ADMIN.'
      x-roles:
      - ADMIN
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AccountResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AccountRequest"
  "/accounts/{accountId}":
    get:
      operationId: getAccount
      summary: Get account detail
      tags:
      - Accounts
      description: 'Allowed roles: ADMIN, ANALYST.'
      x-roles:
      - ADMIN
      - ANALYST
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AccountResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: string
...
  "/accounts/{accountId}/transactions":
    get:
      operationId: listAccountTransactions
      summary: List account transactions
      tags:
      - Accounts
      description: 'Allowed roles: ADMIN, ANALYST, VIEWER.'
      x-roles:
      - ADMIN
      - ANALYST
      - VIEWER
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PageTransactionResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: string
      - name: page
        in: query
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          minimum: 1
          default: 20
      - name: sort
        in: query
        schema:
          type: array
          items:
            type: string
        style: form
        explode: true
        description: Repeat for multiple sorts, e.g. sort=riskScore,desc.
...
  "/alerts":
    get:
      operationId: listAlerts
      summary: List alerts
      tags:
      - Alerts
      description: 'Allowed roles: ADMIN, ANALYST, VIEWER.'
      x-roles:
      - ADMIN
      - ANALYST
      - VIEWER
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PageAlertResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          minimum: 1
          default: 20
      - name: sort
        in: query
        schema:
          type: array
          items:
            type: string
        style: form
        explode: true
        description: Repeat for multiple sorts, e.g. sort=riskScore,desc.
      - name: status
        in: query
        schema:
          "$ref": "#/components/schemas/AlertStatus"
  "/alerts/{alertRef}":
    get:
      operationId: getAlert
      summary: Get alert detail
      tags:
      - Alerts
      description: 'Allowed roles: ADMIN, ANALYST, VIEWER.'
      x-roles:
      - ADMIN
      - ANALYST
      - VIEWER
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AlertResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      parameters:
      - name: alertRef
        in: path
        required: true
        schema:
          type: string
  "/alerts/{alertRef}/status":
    patch:
      operationId: updateAlertStatus
      summary: Update alert status and disposition
      tags:
      - Alerts
      description: ADMIN or ANALYST. Authenticated identity is audited. CLEARED/CLOSED
        alerts and CLOSED/SAR_FILED cases are immutable. A nonblank dispositionReason
        is required to enter a terminal status.
      x-roles:
      - ADMIN
      - ANALYST
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AlertResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      parameters:
      - name: alertRef
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AlertStatusUpdateRequest"
  "/cases":
    get:
      operationId: listCases
      summary: List cases
      tags:
      - Cases
      description: 'Allowed roles: ADMIN, ANALYST, VIEWER.'
      x-roles:
      - ADMIN
      - ANALYST
      - VIEWER
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PageCaseResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          minimum: 1
          default: 20
      - name: sort
        in: query
        schema:
          type: array
          items:
            type: string
        style: form
        explode: true
        description: Repeat for multiple sorts, e.g. sort=riskScore,desc.
      - name: status
        in: query
        schema:
          "$ref": "#/components/schemas/CaseStatus"
    post:
      operationId: createCase
      summary: Create an investigation case
      tags:
      - Cases
      description: 'Allowed roles: ADMIN, ANALYST.'
      x-roles:
      - ADMIN
      - ANALYST
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CaseResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CaseCreateRequest"
  "/cases/{caseRef}":
    get:
      operationId: getCase
      summary: Get case detail
      tags:
      - Cases
      description: 'Allowed roles: ADMIN, ANALYST, VIEWER.'
      x-roles:
      - ADMIN
      - ANALYST
      - VIEWER
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CaseResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      parameters:
      - name: caseRef
        in: path
        required: true
        schema:
          type: string
  "/cases/{caseRef}/status":
    patch:
      operationId: updateCaseStatus
      summary: Update case status and disposition
      tags:
      - Cases
      description: ADMIN or ANALYST. Authenticated identity is audited. CLEARED/CLOSED
        alerts and CLOSED/SAR_FILED cases are immutable. A nonblank dispositionReason
        is required to enter a terminal status.
      x-roles:
      - ADMIN
      - ANALYST
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CaseResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      parameters:
      - name: caseRef
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CaseStatusUpdateRequest"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    IngestionSource:
      type: string
      enum:
      - API
      - BATCH
      - BULK_CSV
      - STREAM
...
    Severity:
      type: string
      enum:
      - LOW
      - MEDIUM
      - HIGH
      - CRITICAL
    CaseStatus:
      type: string
      enum:
      - OPEN
      - INVESTIGATING
      - PENDING_REVIEW
      - ESCALATED
      - SAR_FILED
      - CLOSED
    AccountStatus:
      type: string
      enum:
      - ACTIVE
      - DORMANT
      - CLOSED
      - FROZEN
      - SUSPENDED
    RiskRating:
      type: string
      enum:
      - LOW
      - MEDIUM
      - HIGH
      - CRITICAL
    AlertStatus:
      type: string
      enum:
      - OPEN
      - IN_REVIEW
      - ESCALATED
      - CLEARED
      - CLOSED
...
    TransactionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        transactionRef:
          type: string
        accountId:
          type: string
        customerId:
          type: string
        amount:
          type: number
        currency:
          type: string
        amountBase:
          type: number
        baseCurrency:
          type: string
        transactionType:
          type: string
        direction:
          type: string
        counterpartyName:
          type: string
        counterpartyAccount:
          type: string
        counterpartyCountry:
          type: string
        channel:
          type: string
        jurisdiction:
          type: string
        description:
          type: string
        transactionTime:
          type: string
          format: date-time
        ingestionSource:
          "$ref": "#/components/schemas/IngestionSource"
        createdAt:
          type: string
          format: date-time
...
    AlertStatusUpdateRequest:
      type: object
      properties:
        status:
          "$ref": "#/components/schemas/AlertStatus"
        assignedTo:
          type: string
          maxLength: 100
        disposition:
          type: string
          maxLength: 50
        dispositionReason:
          type: string
          maxLength: 4000
        actor:
          type: string
          description: Deprecated compatibility field. The server ignores this value
            and records the authenticated principal.
          deprecated: true
      required:
      - status
      description: Only ADMIN or ANALYST. Terminal records are immutable. dispositionReason
        is required for terminal transitions; authenticated principal is the audit
        actor.
    AlertResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        alertRef:
          type: string
        customerId:
          type: string
        accountId:
          type: string
        ruleCode:
          type: string
        triggeredRules:
          type: array
          items:
            type: string
        disposition:
          type: string
        dispositionReason:
          type: string
        title:
          type: string
        explanation:
          type: string
        riskScore:
          type: integer
          format: int32
        severity:
          "$ref": "#/components/schemas/Severity"
        status:
          "$ref": "#/components/schemas/AlertStatus"
        dedupKey:
          type: string
        assignedTo:
          type: string
        transactionRefs:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      description: List views redact customer/account identifiers and narrative fields;
        authorized detail responses contain explanation and dispositions. triggeredRules
        contains all combined rule codes.
    CaseCreateRequest:
      type: object
      properties:
        customerRef:
          type: string
        title:
          type: string
          maxLength: 255
          minLength: 1
        description:
          type: string
        priority:
          "$ref": "#/components/schemas/Severity"
        assignedTo:
          type: string
          maxLength: 100
        alertRefs:
          type: array
          items:
            type: string
        actor:
          type: string
          maxLength: 100
          description: Deprecated compatibility field. The server ignores this value
            and records the authenticated principal.
          deprecated: true
      required:
      - title
    CaseStatusUpdateRequest:
      type: object
      properties:
        status:
          "$ref": "#/components/schemas/CaseStatus"
        disposition:
          type: string
          maxLength: 50
        dispositionReason:
          type: string
          maxLength: 4000
        assignedTo:
          type: string
          maxLength: 100
        actor:
          type: string
          description: Deprecated compatibility field. The server ignores this value
            and records the authenticated principal.
          deprecated: true
      required:
      - status
      description: Only ADMIN or ANALYST. Terminal records are immutable. dispositionReason
        is required for terminal transitions; authenticated principal is the audit
        actor.
    CaseResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        caseRef:
          type: string
        customerId:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          "$ref": "#/components/schemas/CaseStatus"
        priority:
          "$ref": "#/components/schemas/Severity"
        assignedTo:
          type: string
        disposition:
          type: string
        dispositionReason:
          type: string
        alertRefs:
          type: array
          items:
            type: string
        openedAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AccountResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        accountId:
          type: string
        customerId:
          type: string
        accountType:
          type: string
        riskRating:
          "$ref": "#/components/schemas/RiskRating"
        accountStatus:
          "$ref": "#/components/schemas/AccountStatus"
        currency:
          type: string
        openDate:
          type: string
          format: date
        closeDate:
          type: string
          format: date
        branchCode:
          type: string
        branchCity:
          type: string
        currentBalance:
          type: number
        avgMonthlyBalance6m:
          type: number
        creditLimit:
          type: number
        creditUtilizationPct:
          type: number
        overdraftEnabled:
          type: boolean
        cardType:
          type: string
        jointAccount:
          type: boolean
        numLinkedDevices:
          type: integer
          format: int32
        mobileBankingEnrolled:
          type: boolean
        lastLoginDate:
          type: string
          format: date
        avgMonthlyTxnCount:
          type: integer
          format: int32
        accountTier:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AccountRequest:
      type: object
      properties:
        accountId:
          type: string
          maxLength: 64
          minLength: 1
        customerId:
          type: string
          maxLength: 64
          minLength: 1
        accountType:
          type: string
          maxLength: 30
        riskRating:
          "$ref": "#/components/schemas/RiskRating"
        accountStatus:
          "$ref": "#/components/schemas/AccountStatus"
        currency:
          type: string
          minLength: 1
          pattern: "^[A-Za-z]{3}$"
        openDate:
          type: string
          format: date
        closeDate:
          type: string
          format: date
        branchCode:
          type: string
          maxLength: 30
        branchCity:
          type: string
          maxLength: 100
        currentBalance:
          type: number
        avgMonthlyBalance6m:
          type: number
        creditLimit:
          type: number
        creditUtilizationPct:
          type: number
        overdraftEnabled:
          type: boolean
        cardType:
          type: string
          maxLength: 30
        jointAccount:
          type: boolean
        numLinkedDevices:
          type: integer
          format: int32
        mobileBankingEnrolled:
          type: boolean
        lastLoginDate:
          type: string
          format: date
        avgMonthlyTxnCount:
          type: integer
          format: int32
        accountTier:
          type: string
          maxLength: 30
      required:
      - accountId
      - customerId
      - currency
...
    PageAccountResponse:
      type: object
      required:
      - content
      - page
      - size
      - totalElements
      - totalPages
      - last
      properties:
        content:
          type: array
          items:
            "$ref": "#/components/schemas/AccountResponse"
        page:
          type: integer
          minimum: 0
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        last:
          type: boolean
    PageTransactionResponse:
      type: object
      required:
      - content
      - page
      - size
      - totalElements
      - totalPages
      - last
      properties:
        content:
          type: array
          items:
            "$ref": "#/components/schemas/TransactionResponse"
        page:
          type: integer
          minimum: 0
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        last:
          type: boolean
    PageAlertResponse:
      type: object
      required:
      - content
      - page
      - size
      - totalElements
      - totalPages
      - last
      properties:
        content:
          type: array
          items:
            "$ref": "#/components/schemas/AlertResponse"
        page:
          type: integer
          minimum: 0
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        last:
          type: boolean
    PageCaseResponse:
      type: object
      required:
      - content
      - page
      - size
      - totalElements
      - totalPages
      - last
      properties:
        content:
          type: array
          items:
            "$ref": "#/components/schemas/CaseResponse"
        page:
          type: integer
          minimum: 0
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        last:
          type: boolean
...
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        timestamp:
          type: string
          format: date-time
        fieldErrors:
          type: object
          additionalProperties:
            type: string
        errorType:
          type: string
        sourceReference:
          type: string
  responses:
    Unauthorized:
      description: Missing or invalid Basic credentials
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    Forbidden:
      description: Authenticated role cannot access this resource
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    Problem:
      description: Request failure; RFC 9457 problem details, including 400 validation,
        404 missing resource, 409 duplicate/conflict, 422 missing parent, or 500 unexpected
        failure.
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"